// src/server/shared/lib/pncp/models/RecuperarDadosDocumentoTermoContratoDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RecuperarDadosDocumentoTermoContratoDTO = {
  uri?: string;
  id?: number;
  url?: string;
  cnpj?: string;
  anoCompra?: number;
  sequencialCompra?: number;
  dataPublicacaoPncp?: string;
  titulo?: string;
  sequencialDocumento?: number;
  tipoDocumentoNome?: string;
};